add rudyredhat1 to owner list (#678)
Hi Team,

I need permission to close all the SA migration PRs and then delete the
respective branch with that. So can anyone help me to merge this? And
also checking if these changes will help me for the respective
modification in the branch or not.

cc @creydr @Kaustubh-pande @dsimansk 

Thanks